Add 103-fibonacci.c

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,9 +8,9 @@
 int main(void)
 {
 	int fn;
-	long previous;
-	long current;
-	long next;
+	int previous;
+	int current;
+	int next;
 	int i;
 	int sum;
